+++ e2e-polybft/txpool_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/0xPolygon/polygon-edge/e2ev3/framework"
+	"github.com/0xPolygon/polygon-edge/e2e-polybft/framework"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
